requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# LLM integration (choose one)
# openai>=1.0.0
anthropic>=0.18.0

# Templating
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
